requirements.txt
snowflake-snowpark-python
pandas
shortuuid